backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
PyMuPDF
pdf2image
reportlab